# libs/domains/framework/backend/feature-agent-controller/spec/openapi.yaml
openapi: 3.1.0
info:
  title: Agent Controller API
  version: 1.0.0
  description: HTTP API for managing clients and proxied agent operations
  contact:
    name: ForePath
    email: hi@forepath.io
    url: https://forepath.io
servers:
  - url: http://localhost:3100/api
    description: Local development
security:
  - bearerAuth: []
paths:
...
  /clients/{id}/agents/{agentId}/autonomy:
    get:
      summary: Get per-agent autonomy settings
      operationId: getClientAgentAutonomy
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: agentId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Autonomy configuration
    put:
      summary: Create or update per-agent autonomy settings
      description: |
        Requires workspace management rights (see ClientResponseDto.canManageWorkspaceConfiguration).
        Interactive user context required; API key auth may manage when applicable.
      operationId: upsertClientAgentAutonomy
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: agentId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Saved autonomy configuration
        '403':
          description: |
            No client access, or workspace management rights required (plain client member cannot update autonomy).
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...